Added mobile responsiveness and comparison section
Modified the CSS so the web app is more mobile responsive, including changing the "scrollTo" buttons so that they are just icons, are smaller, and when it's a mobile device, they will shift position to a row at the top of the page. However, this page is just one long column now. Also added a comparison section and button so users can basically create their own sub-list of items from the search results, then compare them together without having to scroll through each of the 3 API's massive lists.